Create EnsembleMixin
