+++ etna/ensembles/__init__.py
@@ -1,2 +1,3 @@
+from etna.ensembles.base import EnsembleMixin
 from etna.ensembles.stacking_ensemble import StackingEnsemble
 from etna.ensembles.voting_ensemble import VotingEnsemble
